odds training service has been created

+++ odds-training/src/main/java/com/poorbet/oddstraining/model/TeamPower.java
@@ -3,7 +3,7 @@ package com.poorbet.oddstraining.model;
 import org.hibernate.validator.constraints.UUID;
 
 public record TeamPower(
-        double attackPower,
-        double defencePower
+        int attackPower,
+        int defencePower
 ) {}
